src/third_party/blink/web_tests/external/wpt/css/css-values/minmax-number-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'min(.1)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'calc(0.1)' should round-trip exactly in specified values. expected "calc(0.1)" but got "0.1"
FAIL 'scale(min(.1))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(calc(0.1))' should round-trip exactly in specified values. expected "scale(calc(0.1))" but got "scale(0.1)"
PASS 'min(.1)' as a computed value should serialize as '0.1'.
PASS 'scale(min(.1))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'max(.1)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'calc(0.1)' should round-trip exactly in specified values. expected "calc(0.1)" but got "0.1"
FAIL 'scale(max(.1))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(calc(0.1))' should round-trip exactly in specified values. expected "scale(calc(0.1))" but got "scale(0.1)"
PASS 'max(.1)' as a computed value should serialize as '0.1'.
PASS 'scale(max(.1))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'min(.1, .2, .3)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'calc(0.1)' should round-trip exactly in specified values. expected "calc(0.1)" but got "0.1"
FAIL 'scale(min(.1, .2, .3))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(calc(0.1))' should round-trip exactly in specified values. expected "scale(calc(0.1))" but got "scale(0.1)"
PASS 'min(.1, .2, .3)' as a computed value should serialize as '0.1'.
PASS 'scale(min(.1, .2, .3))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'max(.1, .2, .3)' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'calc(0.3)' should round-trip exactly in specified values. expected "calc(0.3)" but got "0.3"
FAIL 'scale(max(.1, .2, .3))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(calc(0.3))' should round-trip exactly in specified values. expected "scale(calc(0.3))" but got "scale(0.3)"
PASS 'max(.1, .2, .3)' as a computed value should serialize as '0.3'.
PASS 'scale(max(.1, .2, .3))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'min(.3, .2, .1)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'calc(0.1)' should round-trip exactly in specified values. expected "calc(0.1)" but got "0.1"
FAIL 'scale(min(.3, .2, .1))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(calc(0.1))' should round-trip exactly in specified values. expected "scale(calc(0.1))" but got "scale(0.1)"
PASS 'min(.3, .2, .1)' as a computed value should serialize as '0.1'.
PASS 'scale(min(.3, .2, .1))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'max(.3, .2, .1)' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'calc(0.3)' should round-trip exactly in specified values. expected "calc(0.3)" but got "0.3"
FAIL 'scale(max(.3, .2, .1))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(calc(0.3))' should round-trip exactly in specified values. expected "scale(calc(0.3))" but got "scale(0.3)"
PASS 'max(.3, .2, .1)' as a computed value should serialize as '0.3'.
PASS 'scale(max(.3, .2, .1))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'calc(min(.1) + min(.2))' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'calc(0.3)' should round-trip exactly in specified values. expected "calc(0.3)" but got "0.3"
FAIL 'scale(calc(min(.1) + min(.2)))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(calc(0.3))' should round-trip exactly in specified values. expected "scale(calc(0.3))" but got "scale(0.3)"
PASS 'calc(min(.1) + min(.2))' as a computed value should serialize as '0.3'.
PASS 'scale(calc(min(.1) + min(.2)))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'calc(max(.1) + max(.2))' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'calc(0.3)' should round-trip exactly in specified values. expected "calc(0.3)" but got "0.3"
FAIL 'scale(calc(max(.1) + max(.2)))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(calc(0.3))' should round-trip exactly in specified values. expected "scale(calc(0.3))" but got "scale(0.3)"
PASS 'calc(max(.1) + max(.2))' as a computed value should serialize as '0.3'.
PASS 'scale(calc(max(.1) + max(.2)))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'calc(.1 + min(.1))' as a specified value should serialize as 'calc(0.2)'. assert_equals: 'calc(0.2)' should round-trip exactly in specified values. expected "calc(0.2)" but got "0.2"
FAIL 'scale(calc(.1 + min(.1)))' as a specified value should serialize as 'scale(calc(0.2))'. assert_equals: 'scale(calc(0.2))' should round-trip exactly in specified values. expected "scale(calc(0.2))" but got "scale(0.2)"
PASS 'calc(.1 + min(.1))' as a computed value should serialize as '0.2'.
PASS 'scale(calc(.1 + min(.1)))' as a computed value should serialize as 'matrix(0.2, 0, 0, 0.2, 0, 0)'.
FAIL 'calc(max(.1) + .1)' as a specified value should serialize as 'calc(0.2)'. assert_equals: 'calc(0.2)' should round-trip exactly in specified values. expected "calc(0.2)" but got "0.2"
FAIL 'scale(calc(max(.1) + .1))' as a specified value should serialize as 'scale(calc(0.2))'. assert_equals: 'scale(calc(0.2))' should round-trip exactly in specified values. expected "scale(calc(0.2))" but got "scale(0.2)"
PASS 'calc(max(.1) + .1)' as a computed value should serialize as '0.2'.
PASS 'scale(calc(max(.1) + .1))' as a computed value should serialize as 'matrix(0.2, 0, 0, 0.2, 0, 0)'.
Harness: the test ran to completion.
